AlertDialog class and NoteList TextWrapping
_

Added an AlertDialog class to handle displaying messages to the user. TextWrapping has now been added to Notes List in the Source Viewing page, to present the note content in a more orderly fashion. Note Input Box allows for multiline text, which allows the user to paste copied text that spans for multiple lines.